requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
